refactor: remove ENSNODE_PUBLIC_URL, ENSADMIN_URL, PORT env variables
