+++ apps/ensadmin/src/components/recent-registrations/components.tsx
@@ -8,6 +8,8 @@ import {
 import { Badge } from "@/components/ui/badge";
 import { Button } from "@/components/ui/button";
 import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card";
+import { useActiveConnection } from "@/hooks/active/use-active-connection";
+import { useSelectedConnection } from "@/hooks/active/use-selected-connection";
 import { useRawConnectionUrlParam } from "@/hooks/use-connection-url-param";
 import { cn } from "@/lib/utils";
 import type { ENSNamespaceId } from "@ensnode/datasources";
@@ -19,7 +21,7 @@ import {
 } from "@ensnode/ensnode-sdk";
 import { useAutoAnimate } from "@formkit/auto-animate/react";
 import Link from "next/link";
-import { useEffect, useState } from "react";
+import { useEffect, useMemo, useState } from "react";
 import { useRecentRegistrations } from "./hooks";
 
 /**
@@ -108,7 +110,7 @@ export function RecentRegistrations({
     );
   }
 
-  const { ensNodePublicUrl: ensNodeUrl, namespace: namespaceId } = ensIndexerConfig;
+  const { namespace: namespaceId } = ensIndexerConfig;
 
   return (
     <Card className="w-full">
@@ -118,20 +120,13 @@ export function RecentRegistrations({
         </CardTitle>
       </CardHeader>
       <CardContent>
-        {isClient && (
-          <RegistrationsList
-            ensNodeUrl={ensNodeUrl}
-            namespaceId={namespaceId}
-            maxRecords={maxRecords}
-          />
-        )}
+        {isClient && <RegistrationsList namespaceId={namespaceId} maxRecords={maxRecords} />}
       </CardContent>
     </Card>
   );
 }
 
 interface RegistrationsListProps {
-  ensNodeUrl: URL;
   namespaceId: ENSNamespaceId;
   maxRecords: number;
 }
@@ -142,9 +137,10 @@ interface RegistrationsListProps {
  * @param ensNodeMetadata data about connected ENSNode instance necessary for fetching registrations
  * @param ensNodeUrl URL of currently selected ENSNode instance
  */
-function RegistrationsList({ ensNodeUrl, namespaceId, maxRecords }: RegistrationsListProps) {
+function RegistrationsList({ namespaceId, maxRecords }: RegistrationsListProps) {
+  const { rawSelectedConnection } = useSelectedConnection();
   const recentRegistrationsQuery = useRecentRegistrations({
-    ensNodeUrl,
+    ensNodeUrl: useMemo(() => new URL(rawSelectedConnection), [rawSelectedConnection]),
     namespaceId,
     maxRecords,
   });
