+++ apps/ensindexer/.env.local.example
@@ -1,5 +1,5 @@
 # The port ENSIndexer listens on for HTTP requests.
-# Optional. If this is not set, the default value is set to `DEFAULT_PORT` (42069).
+# Optional. If this is not set, the default value is set to Ponder's default of 42069.
 PORT=42069
 
 # RPC configuration
@@ -177,12 +177,6 @@ LABEL_SET_ID=subgraph
 # and LABEL_SET_VERSION must be set to 0.
 LABEL_SET_VERSION=0
 
-# The ENSIndexer public service URL for handling external API requests
-# Required. When the root route `/` of ENSIndexer receives a request, ENSIndexer redirects to
-# the configured ENSADMIN_URL with an instruction for that ENSAdmin instance to connect back
-# to this provided URL for querying state about the ENSNode instance.
-ENSNODE_PUBLIC_URL=http://localhost:42069
-
 # The "primary" ENSIndexer service URL
 # Required. This must be an instance of ENSIndexer using either `ponder start`
 # or `ponder dev`, and not `ponder serve`.
@@ -196,15 +190,6 @@ ENSNODE_PUBLIC_URL=http://localhost:42069
 # the related ENSIndexer instance started using `ponder start` or
 # `ponder dev` that is writing to the same ENSDb.
 ENSINDEXER_URL=http://localhost:42069
-
-# The ENSAdmin service URL
-# Optional. When the root route `/` of ENSIndexer receives a request, ENSIndexer redirects
-# to this provided ENSAdmin URL with an instruction for that ENSAdmin instance to connect
-# back to the configured ENSNODE_PUBLIC_URL.
-#
-# If this is not set, DEFAULT_ENSADMIN_URL (https://admin.ensnode.io) will be used
-# to provide easy access to an ENSAdmin UI.
-ENSADMIN_URL=https://admin.ensnode.io
 
 # A feature flag to enable/disable ENSIndexer's Subgraph Compatible Indexing Behavior
 # Optional. If this is not set, the default value is set to `DEFAULT_SUBGRAPH_COMPAT` (false).
