+++ apps/ensindexer/src/lib/lib-config.ts
@@ -1,8 +1,6 @@
 import type { RpcConfigEnvironment } from "@/config/types";
 import { type ChainIdString, deserializeChainId } from "@ensnode/ensnode-sdk";
 
-export const DEFAULT_ENSADMIN_URL = new URL("https://admin.ensnode.io");
-export const DEFAULT_PORT = 42069;
 export const DEFAULT_SUBGRAPH_COMPAT = false;
 
 /**
