+++ packages/ensnode-sdk/src/ensindexer/config/serialized-types.ts
@@ -1,4 +1,4 @@
-import type { ChainId, UrlString } from "../../shared";
+import type { ChainId } from "../../shared";
 import type { ENSIndexerPublicConfig } from "./types";
 
 export type SerializedIndexedChainIds = Array<ChainId>;
@@ -7,17 +7,7 @@ export type SerializedIndexedChainIds = Array<ChainId>;
  * Serialized representation of {@link ENSIndexerPublicConfig}
  */
 export interface SerializedENSIndexerPublicConfig
-  extends Omit<ENSIndexerPublicConfig, "ensAdminUrl" | "ensNodePublicUrl" | "indexedChainIds"> {
-  /**
-   * Serialized representation of {@link ENSIndexerPublicConfig.ensAdminUrl}.
-   */
-  ensAdminUrl: UrlString;
-
-  /**
-   * Serialized representation of {@link ENSIndexerPublicConfig.ensNodePublicUrl}.
-   */
-  ensNodePublicUrl: UrlString;
-
+  extends Omit<ENSIndexerPublicConfig, "indexedChainIds"> {
   /**
    * Array representation of {@link ENSIndexerPublicConfig.indexedChainIds}.
    */
